backend/requirements.txt
fastapi
uvicorn
motor
pymongo
python-jose[cryptography]
passlib
python-multipart
email-validator